Refactoring & Fixes (#88)
* Added Multi30k multimodal MT test set metadata
* Refactored all tokenizers into respective classes (fixes #85)
* Refactored all metrics into respective classes
* Moved utility functions into utils.py
* Implemented signatures using BLEUSignature and CHRFSignature classes, expose `Signature().info`
* metrics: Signature().info is now exposed (fixes #75)
* Simplified checking of Chinese characters (fixes #5)
* Unified common regexp tokenization codes for tokenizers (fixes #27)
* Fixed --detail failing when no test sets are provided
* Fixed multi-reference BLEU failing when tab-delimited reference stream is used
* Removed lowercase option for ChrF which was not functional (#85)
* Simplified ChrF and used the same I/O logic as BLEU to allow for future
   multi-reference reading
* Added score regression tests for chrF using reference chrF++ implementation
* Added multi-reference & tokenizer & signature tests
* Pin mecab version to 0.996.5 as the newer ones are incompatible (fixes #94)
* bump version to 1.4.11
