+++ setup.py
@@ -129,7 +129,7 @@ setup(
     # your project is installed. For an analysis of "install_requires" vs pip's
     # requirements files see:
     # https://packaging.python.org/en/latest/requirements.html
-    install_requires = ['typing;python_version<"3.5"', 'portalocker'],
+    install_requires = ['typing;python_version<"3.5"', 'portalocker', 'mecab-python3==0.996.5'],
 
     # List additional groups of dependencies here (e.g. development
     # dependencies). You can install these using the following syntax,
